Showing Jotham

+++ common/src/main/java/com/tendai/common/data/model/Artist.kt
@@ -2,7 +2,7 @@ package com.tendai.common.data.model
 
 
 data class Artist(
-var artistId: Long,
+    var artistId: Int,
     var artistName: String,
     var numberOfAlbums: Int,
     var numberOfTracks: Int
